v8 cache

+++ packages/parcel-plugin-v8-cache/JSPackager.js
@@ -1,7 +1,8 @@
 const fs = require('fs-extra')
 const ParcelJSPackager = global.JSPackager || require('parcel/lib/packagers/JSPackager')
+const { promisify } = require('util')
 const { minify } = require('terser')
-const { relative, join, dirname } = require('path')
+const { relative, join, dirname, basename } = require('path')
 
 const EXCLUDES = ['downloadPage']
 
@@ -12,15 +13,18 @@ global.JSPackager = module.exports = class JSPackager extends ParcelJSPackager {
     this.v8CacheFile = join(b.options.outDir, 'v8-cache.js')
   }
   async setup () {
-    const data = await fs.readFile(require.resolve('v8-compile-cache'))
-    await fs.writeFile(this.v8CacheFile, minify(data.toString(), OPTIONS).code)
     await super.setup()
-  }
-  async start () {
-    if (EXCLUDES.every(it => !this.bundle.name.includes(it))) {
-      const file = relative(dirname(this.bundle.name), this.v8CacheFile).replace(/\\/g, '/')
-      await this.write(`typeof require!='undefined'&&require('./${file}');`)
+    if (!await fs.pathExists(this.v8CacheFile)) {
+      const data = await fs.readFile(require.resolve('v8-compile-cache'))
+      await fs.writeFile(this.v8CacheFile, minify(data.toString(), OPTIONS).code)
     }
-    await super.start()
+    if (EXCLUDES.every(it => !this.bundle.name.includes(it))) {
+      const file = this.bundle.name.replace(/js$/, 'c.js')
+      const cFile = relative(dirname(this.bundle.name), this.v8CacheFile).replace(/\\/g, '/')
+      this.dest.end(`require('./${cFile}');require('./${basename(file)}')`)
+      this.dest = fs.createWriteStream(file)
+      this.dest.write = promisify(this.dest.write.bind(this.dest))
+      this.dest.end = promisify(this.dest.end.bind(this.dest))
+    }
   }
 }
